python/requirements.txt
mlx
numpy
Pillow
safetensors
torch  # Required for weight conversion, optional for inference
